Add control utility
This changes the stored RSA secret (it's now just wrapkey, instead of nonce | wrapkey).
Instead of that, a nonce is generated by using the ring NonceSequence structure
based on a counter, to ensure that it's possible to store secrets that are more
than a single GCM block long.

This means any previously encrypted secrets can't be read anymore.

Signed-off-by: Patrick Uiterwijk <patrick@puiterwijk.org>